feat: Adjusted Widgets
