+++ lib/src/widgets/widgets_export.dart
@@ -1,3 +1,3 @@
 export 'texts/responsive_text_widget.dart';
-export 'fields/app_field_widget.dart';
+export 'fields/app_web_field_widget.dart';
 export 'buttons/social_circle_button_widget.dart';
